+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.10](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.9.9...@wundergraph/nextjs@0.9.10) (2023-03-29)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.9.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.9.8...@wundergraph/nextjs@0.9.9) (2023-03-28)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.9.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.9.7...@wundergraph/nextjs@0.9.8) (2023-03-27)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.9.7](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.9.6...@wundergraph/nextjs@0.9.7) (2023-03-23)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
 ## [0.9.6](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.9.5...@wundergraph/nextjs@0.9.6) (2023-03-16)
 
 **Note:** Version bump only for package @wundergraph/nextjs